factor out common unification infrastructure into new module
