+++ src/swarm-lang/Swarm/Effect/Unify/Naive.hs
@@ -7,7 +7,11 @@
 -- SPDX-License-Identifier: BSD-3-Clause
 --
 -- Description: Naive (slow) substitution-based implementation of
--- unification.
+-- unification.  Uses a simple but expensive-to-maintain invariant on
+-- substitutions, and returns a substitution from unification which
+-- must then be composed with the substitution being tracked.
+--
+-- Not used in Swarm, but useful for testing/comparison.
 module Swarm.Effect.Unify.Naive where
 
 import Control.Algebra
@@ -20,82 +24,17 @@ import Control.Effect.Throw (Throw, throwError)
 import Control.Monad.Free
 import Control.Monad.Trans (MonadIO)
 import Data.Function (on)
-import Data.Map (Map, (!?))
+import Data.Map ((!?))
 import Data.Map qualified as M
 import Data.Map.Merge.Lazy qualified as M
 import Data.Maybe (fromMaybe)
-import Data.Set (Set)
 import Data.Set qualified as S
 import Swarm.Effect.Unify
+import Swarm.Effect.Unify.Common
 import Swarm.Language.Types hiding (Type)
 
 ------------------------------------------------------------
--- Naive carrier
-
-newtype UnificationC m a = UnificationC
-  {unUnificationC :: StateC (Subst IntVar UType) (StateC FreshVarCounter (ThrowC UnificationError m)) a}
-  deriving newtype (Functor, Applicative, Alternative, Monad, MonadIO)
-
-newtype FreshVarCounter = FreshVarCounter {getFreshVarCounter :: Int}
-  deriving (Eq, Ord, Enum)
-
-instance Algebra sig m => Algebra (Unification :+: sig) (UnificationC m) where
-  alg hdl sig ctx = UnificationC $ case sig of
-    L (Unify t1 t2) -> do
-      s1 <- get @(Subst IntVar UType)
-      let t1' = subst s1 t1
-          t2' = subst s1 t2
-      s2 <- unify t1' t2'
-      modify (s2 @@)
-      return $ (subst s2 t1' <$ ctx)
-    L (ApplyBindings t) -> do
-      s <- get @(Subst IntVar UType)
-      return $ subst s t <$ ctx
-    L FreshIntVar -> do
-      v <- IntVar <$> gets getFreshVarCounter
-      modify @FreshVarCounter succ
-      return $ v <$ ctx
-    L (FreeUVars t) -> do
-      s <- get @(Subst IntVar UType)
-      return $ fuvs (subst s t) <$ ctx
-    R other -> alg (unUnificationC . hdl) (R (R (R other))) ctx
-
-runUnification :: Algebra sig m => UnificationC m a -> m (Either UnificationError a)
-runUnification = unUnificationC >>> evalState idS >>> evalState (FreshVarCounter 0) >>> runThrow
-
--- | A value of type @Subst n a@ is a substitution which maps
---   names of type @n@ (the /domain/, see 'dom') to values of type
---   @a@.  Substitutions can be /applied/ to certain terms (see
---   'subst'), replacing any free occurrences of names in the
---   domain with their corresponding values.  Thus, substitutions can
---   be thought of as functions of type @Term -> Term@ (for suitable
---   @Term@s that contain names and values of the right type).
---
---   Concretely, substitutions are stored using a @Map@.
---
---   As an invariant, map keys never show up in any of the values.
---   For example, we could have  @{x -> a+5, y -> 5}@  but not
---   @{x -> a+y, y -> 5}@.
-newtype Subst n a = Subst {getSubst :: Map n a}
-  deriving (Eq, Ord, Show)
-
-instance Functor (Subst n) where
-  fmap f (Subst m) = Subst (M.map f $ m)
-
--- | The domain of a substitution is the set of names for which the
---   substitution is defined.
-dom :: Subst n a -> Set n
-dom = M.keysSet . getSubst
-
--- | The identity substitution, /i.e./ the unique substitution with an
---   empty domain, which acts as the identity function on terms.
-idS :: Subst n a
-idS = Subst M.empty
-
--- | Construct a singleton substitution, which maps the given name to
---   the given value.
-(|->) :: n -> a -> Subst n a
-x |-> t = Subst (M.singleton x t)
+-- Substitutions
 
 -- | Class of things supporting substitution.  @Substitutes n b a@ means
 --   that we can apply a substitution of type @Subst n b@ to a
@@ -125,21 +64,64 @@ instance (Show n, Ord n, Functor f) => Substitutes n (Free f n) (Free f n) where
 compose :: (Ord n, Substitutes n a a, Foldable t) => t (Subst n a) -> Subst n a
 compose = foldr (@@) idS
 
--- | Create a substitution from an association list of names and
---   values.
-fromList :: Ord n => [(n, a)] -> Subst n a
-fromList = Subst . M.fromList
+------------------------------------------------------------
+-- Carrier type
 
--- | Convert a substitution into an association list.
-toList :: Subst n a -> [(n, a)]
-toList = M.assocs . getSubst
+-- Note: this carrier type and the runUnification function are
+-- identical between this module and Swarm.Effect.Unify.Fast, but it
+-- seemed best to duplicate it, so we can modify the carriers
+-- independently in the future if we want.
 
--- | Look up the value a particular name maps to under the given
---   substitution; or return @Nothing@ if the name being looked up is
---   not in the domain.
-lookup :: Ord n => n -> Subst n a -> Maybe a
-lookup x (Subst m) = M.lookup x m
+-- | Carrier type for unification: we maintain a current substitution,
+--   a counter for generating fresh unification variables, and can
+--   throw unification errors.
+newtype UnificationC m a = UnificationC
+  { unUnificationC ::
+      StateC (Subst IntVar UType) (StateC FreshVarCounter (ThrowC UnificationError m)) a
+  }
+  deriving newtype (Functor, Applicative, Alternative, Monad, MonadIO)
 
+-- | Counter for generating fresh unification variables.
+newtype FreshVarCounter = FreshVarCounter {getFreshVarCounter :: Int}
+  deriving (Eq, Ord, Enum)
+
+-- | Run a 'Unification' effect via the 'UnificationC' carrier.
+runUnification :: Algebra sig m => UnificationC m a -> m (Either UnificationError a)
+runUnification =
+  unUnificationC >>> evalState idS >>> evalState (FreshVarCounter 0) >>> runThrow
+
+------------------------------------------------------------
+-- Unification
+
+-- | Naive implementation of the 'Unification' effect in terms of the
+--   'UnificationC' carrier.
+--
+--   We maintain an invariant on the current @Subst@ that map keys
+--   never show up in any of the values.  For example, we could have
+--   @{x -> a+5, y -> 5}@ but not @{x -> a+y, y -> 5}@.
+instance Algebra sig m => Algebra (Unification :+: sig) (UnificationC m) where
+  alg hdl sig ctx = UnificationC $ case sig of
+    L (Unify t1 t2) -> do
+      s1 <- get @(Subst IntVar UType)
+      let t1' = subst s1 t1
+          t2' = subst s1 t2
+      s2 <- unify t1' t2'
+      modify (s2 @@)
+      return $ (subst s2 t1' <$ ctx)
+    L (ApplyBindings t) -> do
+      s <- get @(Subst IntVar UType)
+      return $ subst s t <$ ctx
+    L FreshIntVar -> do
+      v <- IntVar <$> gets getFreshVarCounter
+      modify @FreshVarCounter succ
+      return $ v <$ ctx
+    L (FreeUVars t) -> do
+      s <- get @(Subst IntVar UType)
+      return $ fuvs (subst s t) <$ ctx
+    R other -> alg (unUnificationC . hdl) (R (R (R other))) ctx
+
+-- | Unify two types and return the mgu, i.e. the smallest
+--   substitution which makes them equal.
 unify ::
   Has (Throw UnificationError) sig m =>
   UType ->
@@ -157,6 +139,8 @@ unify ty1 ty2 = case (ty1, ty2) of
     | otherwise -> return $ x |-> y
   (Free t1, Free t2) -> unifyF t1 t2
 
+-- | Unify two non-variable terms and return an mgu, i.e. the smallest
+--   substitution which makes them equal.
 unifyF ::
   Has (Throw UnificationError) sig m =>
   TypeF UType ->
